chore: clean up main.rs and simulation-mode
